Move around socket simulator code
Separate memcached-specific stuff from sockets_simulator.c.

Also make separate dir for custom socket simulators.
