+++ runtime/Makefile.cmake.bitcode.rules
@@ -81,14 +81,14 @@ endif
 # Compute the directory to put build files
 LOCAL_BUILD_DIR := $(abspath $(ROOT_OBJ)/$(DIR_SUFFIX))
 
-C_SRCS := $(sort $(shell echo $(SRC_DIR)/*.c))
+C_SRCS := $(foreach subdir,$(SRC_DIRS),$(sort $(shell echo $(SRC_DIR)/$(subdir)/*.c)))
 
 # Sanity check: Make sure at least one source file was found
 ifneq ($(strip $(filter %*.c,$(C_SRCS))),)
 $(error Failed to find C source files in $(SRC_DIR))
 endif
 
-C_SRCS_NO_DIR := $(notdir $(C_SRCS))
+C_SRCS_NO_DIR := $(subst $(SRC_DIR)/,,$(C_SRCS))
 BC_FILES_NO_DIR := $(C_SRCS_NO_DIR:.c=.bc)
 BC_FILES := $(addprefix $(LOCAL_BUILD_DIR)/,$(BC_FILES_NO_DIR))
 
@@ -114,7 +114,7 @@ $(BC_FILES) : $(ADDITIONAL_BUILD_DEPS) $(LLVMCC_FLAGS_FILE)
 # Pattern rule to build bitcode files
 $(LOCAL_BUILD_DIR)/%.bc : $(SRC_DIR)/%.c
 	@echo "LLVMCC ($(RUNTIME_CONFIG_STRING)) $<"
-	$(Verb) $(MKDIR) -p $(LOCAL_BUILD_DIR)
+	$(Verb) $(MKDIR) -p $(dir $@)
 	$(Verb) $(LLVMCC) $(LLVMCC.Flags) $(LLVMCC.Warnings) $< -c -o $@ -MP -MMD -MF $(LOCAL_BUILD_DIR)/$*.dep
 
 # $(LLVMCC_FLAGS_FILE) depends on `force` which will force the rule to
