+++ runtime/POSIX/sockets.c
@@ -38,7 +38,7 @@
 
 //#include "signals.h"
 #include "netlink.h"
-#include "sockets_simulator.h"
+#include "socksim/sockets_simulator.h"
 
 #include <sys/types.h>
 #include <sys/socket.h>
